+++ lib/k8s/cluster/discovery.ex
@@ -17,17 +17,6 @@ defmodule K8s.Cluster.Discovery do
   @behaviour K8s.Cluster.Discovery.Driver
   @driver Application.get_env(:k8s, :discovery_driver, K8s.Cluster.Discovery.HTTPDriver)
 
-  @typedoc """
-  Resource definition identifier. Format: `{groupVersion, kind, name}`
-
-  Where `kind` is the kubernetes resource type and `name` is the name of the resource, which includes sub-resources.
-
-  E.g.:
-  * `{"apps/v1", "Deployment", "deployments"}
-  * `{"apps/v1", "Deployment", "deployments/status"}
-  """
-  @type resource_definition_identifier_t :: {binary(), binary(), binary()}
-
   @doc """
   Lists Kubernetes `apiVersion`s
 
@@ -55,44 +44,10 @@ defmodule K8s.Cluster.Discovery do
     end
   end
 
-  @doc """
-  Lists identifiers from Kubernetes resource definitions returned from `resource_definitions/2`.
-
-  ## Examples
-      iex> K8s.Cluster.Discovery.resource_identifiers(:test)
-      [{"auditregistration.k8s.io/v1alpha1", "AuditSink", "auditsinks"}, {"settings.k8s.io/v1alpha1", "PodPreset", "podpresets"}, {"apiregistration.k8s.io/v1", "APIService", "apiservices"}]
-  """
-  @spec resource_identifiers(atom(), Keyword.t() | nil) ::
-          {:ok, list(resource_definition_identifier_t)} | {:error, atom()}
-  def resource_identifiers(cluster, opts \\ []) do
-    with {:ok, definitions} <- resource_definitions(cluster, opts) do
-      {:ok, get_identifiers_from_resource_definitions(definitions)}
-    end
-  end
-
-  @spec get_identifiers_from_resource_definitions(list(map())) ::
-          list(resource_definition_identifier_t)
-  defp get_identifiers_from_resource_definitions(groups) do
-    Enum.reduce(groups, [], fn %{"groupVersion" => gv, "resources" => resources}, acc ->
-      resource_identifiers =
-        Enum.map(resources, fn resource ->
-          api_version = resource_api_version(gv, resource)
-          {api_version, resource["kind"], resource["name"]}
-        end)
-
-      acc ++ resource_identifiers
-    end)
-  end
-
   @spec reduce_by_group(list(map())) :: map()
   defp reduce_by_group(groups) do
     Enum.reduce(groups, %{}, fn %{"groupVersion" => gv, "resources" => resources}, acc ->
-      Enum.reduce(resources, acc, fn resource, acc ->
-        api_version = resource_api_version(gv, resource)
-        prev_resources_in_group = Map.get(acc, api_version, [])
-        new_resources_in_group = [resource | prev_resources_in_group]
-        Map.put(acc, api_version, new_resources_in_group)
-      end)
+      Map.put(acc, gv, resources)
     end)
   end