+++ test/k8s/cluster/discovery_test.exs
@@ -13,24 +13,6 @@ defmodule K8s.Cluster.DiscoveryTest do
     end
   end
 
-  describe "resource_identifiers/1" do
-    test "returns a list of resource identifiers" do
-      cluster = :test
-      file = "test/support/discovery/sample_resource_definitions.json"
-      {:ok, resource_identifiers} = Discovery.resource_identifiers(cluster, path: file)
-
-      assert resource_identifiers == [
-               {"batch/v1", "Job", "jobs"},
-               {"apps/v1", "DaemonSet", "daemonsets"},
-               {"apps/v1", "Deployment", "deployments"},
-               {"apps/v1", "Deployment", "deployments/status"},
-               {"v1", "Namespace", "namespaces"},
-               {"v1", "Pod", "pods"},
-               {"policy/v1beta1", "Eviction", "pods/eviction"}
-             ]
-    end
-  end
-
   describe "resource_definitions/1" do
     test "returns full resource definitions" do
       cluster = :test
@@ -49,27 +31,38 @@ defmodule K8s.Cluster.DiscoveryTest do
   end
 
   describe "resources_by_group/2" do
-    test "returns resources group by groupVersion/apiVersion" do
+    test "returns resources group by API version" do
       cluster = :test
       file = "test/support/discovery/sample_resource_definitions.json"
       {:ok, actual} = Discovery.resources_by_group(cluster, path: file)
 
       assert_lists_equal(actual["batch/v1"], [%{"kind" => "Job", "name" => "jobs"}])
+    end
 
-      assert_lists_equal(actual["apps/v1"], [
+    test "includes subresources" do
+      cluster = :test
+      file = "test/support/discovery/sample_resource_definitions.json"
+      {:ok, actual} = Discovery.resources_by_group(cluster, path: file)
+      apps_v1 = actual["apps/v1"]
+      core_v1 = actual["v1"]
+
+      assert_lists_equal(apps_v1, [
         %{"kind" => "DaemonSet", "name" => "daemonsets"},
         %{"kind" => "Deployment", "name" => "deployments"},
         %{"kind" => "Deployment", "name" => "deployments/status"}
       ])
-    end
 
-    test "does not include subresources in their parent groupVersion" do
-      cluster = :test
-      file = "test/support/discovery/sample_resource_definitions.json"
-      {:ok, actual} = Discovery.resources_by_group(cluster, path: file)
-
-      assert_lists_equal(actual["v1"], [
+      assert_lists_equal(core_v1, [
         %{"kind" => "Namespace", "name" => "namespaces"},
+        %{
+          "group" => "policy",
+          "kind" => "Eviction",
+          "name" => "pods/eviction",
+          "namespaced" => true,
+          "singularName" => "",
+          "verbs" => ["create"],
+          "version" => "v1beta1"
+        },
         %{
           "categories" => ["all"],
           "kind" => "Pod",
@@ -87,24 +80,6 @@ defmodule K8s.Cluster.DiscoveryTest do
             "update",
             "watch"
           ]
-        }
-      ])
-    end
-
-    test "groups subresources by _their_ groupVersion" do
-      cluster = :test
-      file = "test/support/discovery/sample_resource_definitions.json"
-      {:ok, actual} = Discovery.resources_by_group(cluster, path: file)
-
-      assert_lists_equal(actual["policy/v1beta1"], [
-        %{
-          "group" => "policy",
-          "kind" => "Eviction",
-          "name" => "pods/eviction",
-          "namespaced" => true,
-          "singularName" => "",
-          "verbs" => ["create"],
-          "version" => "v1beta1"
         }
       ])
     end